[fix] qt type error fixed

+++ app/dashboard/dbusclient.cpp
@@ -22,17 +22,17 @@ DBusClient::~DBusClient() {
 }
 
 
-int DBusClient::speed() const {
-  QDBusReply<int> reply = _iface->call("getSpeed");
-  int value = reply.value();
+qreal DBusClient::speed() const {
+  QDBusReply<qreal> reply = _iface->call("getSpeed");
+  qreal value = reply.value();
 
 
   return value;
 }
 
-int DBusClient::rpm() const {
-  QDBusReply<int> reply = _iface->call("getRpm");
-  int value = reply.value();
+qreal DBusClient::rpm() const {
+  QDBusReply<qreal> reply = _iface->call("getRpm");
+  qreal value = reply.value();
 
   qDebug() << value;
